requirements.txt
alembic==1.12.1
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
